Meetups now appear with title, instead of location (#11)
Small change to r2/templates/meetup.html, exactly the change you would
expect. The issue is that the result is being cached -- in order to see
the change, you must add or edit a meetup, which will clear the cache
(not the browser cache). See r2/controllers/api.py for more detail.