mobile-style: improve mobile display
Rework the mobile display for the site.  The original desktop site was written around the
assumption that all screens were at least 1024px wide, so instead of fighting that, set
the viewport to 1024px and then make tweaks on top of that.  Changes, which will apply to
all screens narrower than 1024px:

* Hide the sidebar, which includes "Getting Started", "Recent Commments", "Recently on EA
  Blogs", and "Top Contributors".

* Stop indenting the body of posts, but do indent the title of posts to leave room for the
  #votes counter.

* Make various things bigger, since we're on a small screen.

This isn't perfect, but it makes it much more usable on mobile.  Primarily, the text is
now big enough to read without zooming.  Some screenshots:

  http://www.jefftk.com/eaf-article-view.png
  http://www.jefftk.com/eaf-comments-view.png
  http://www.jefftk.com/eaf-index-bottom.png
  http://www.jefftk.com/eaf-index-view.png

  (From my phone, a Galaxy S6 with weird DPI settings, so not entirely representative.)

Here's the html of an article page and the index page, with this change patched in:

  http://www.jefftk.com/concerns_with_intentional_insights
  http://www.jefftk.com/ea-index

I haven't tested this pull request except by manually patching my changes into html pages
(the two linked aboved).  That is, I haven't set up a developer environment for the EA
Forum codebase, so there might be something I'm missing about how to integrate my css
change.